requirements.txt
Django==1.10.4
mysqlclient==1.3.9
olefile==0.44
Pillow==4.0.0
django-crispy-forms==1.6.1
django-modeltranslation==0.12
